requirements.txt
pyautogen
openai
playsound
